+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - [#6229](https://github.com/influxdata/influxdb/issues/6229): Fixed aggregate queries with no GROUP BY to include the end time.
 - [#6283](https://github.com/influxdata/influxdb/pull/6283): Fix GROUP BY tag to produce consistent results when a series has no tags.
 - [#3773](https://github.com/influxdata/influxdb/issues/3773): Support empty tags for all WHERE equality operations.
+- [#6270](https://github.com/influxdata/influxdb/issues/6270): tsm1 query engine alloc reduction
 
 ## v0.12.0 [2016-04-05]
 ### Release Notes